+++ examples/typescript/smtp-client-pool/src/types.ts
@@ -1,0 +1,7 @@
+import type Mail from 'nodemailer/lib/mailer/index.js'
+import type SMTPTransport from 'nodemailer/lib/smtp-transport/index.js'
+
+export interface WorkerData {
+  smtpTransport: SMTPTransport.Options
+  mail: Mail.Options
+}