+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add kill handler to worker options allowing to execute custom code when worker is killed.
+- SMTP server pool example: nodemailer.
 
 ## [2.6.25] - 2023-08-13